requirements.txt
streamlit
fastapi
uvicorn
requests
pdfplumber
python-docx
python-dotenv
openai
pytesseract
pdf2image
Pillow
pandas
python-multipart